+++ controlgallery/AtomUIGallery/ShowCases/DataDisplay/ImagePreviewer/Localization/ImagePreviewerShowCaseLangResourceKind.cs
@@ -44,7 +44,6 @@ public enum ImagePreviewerShowCaseLangResourceKind
     SemanticPopupBodyDescription,
     SemanticPopupFooterDescription,
     SemanticPopupActionsDescription,
-    SemanticPopupCloseDescription,
     SemanticPartStyleTitle,
     SemanticPartStyleDescription
 }